DumpedPrivateKeyTest: move p2shAddressCreationFromKeys from LegacyAddressTest
This test has dependencies on o.b.crypto and is arguably a test of
DumpedPrivateKey.
